+++ OsmAnd/src/net/osmand/plus/mapcontextmenu/MenuBuilder.java
@@ -49,8 +49,8 @@ import net.osmand.plus.mapcontextmenu.builders.cards.ImageCard.GetImageCardsTask
 import net.osmand.plus.mapcontextmenu.builders.cards.NoImagesCard;
 import net.osmand.plus.mapcontextmenu.controllers.TransportStopController;
 import net.osmand.plus.osmedit.opr.OPRWebviewActivity;
-import net.osmand.plus.osmedit.utils.SecUtils;
-import net.osmand.plus.osmedit.utils.util.exception.FailedVerificationException;
+import net.osmand.plus.osmedit.utils.opendb.SecUtils;
+import net.osmand.plus.osmedit.utils.opendb.util.exception.FailedVerificationException;
 import net.osmand.plus.poi.PoiUIFilter;
 import net.osmand.plus.render.RenderingIcons;
 import net.osmand.plus.transport.TransportStopRoute;
